Started adding functions to Trace and TraceData
Got around previous problem by deciding that you cannot just add two
tracedata objects together.  You must instead add the data from one to
the other, which explicitly determines which metadata you get.
if A and B are Trace or TraceData objects... then
X = A + B.data  returns an object of type A, with A's metadata.